+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "@sendgrid/mail": "^7.6.2",
     "bcrypt": "^5.0.1",
+    "bcryptjs": "^2.4.3",
     "cloudinary": "^1.29.0",
     "cors": "^2.8.5",
     "datauri": "^4.1.0",
