+++ routes/users/Auth.js
@@ -1,0 +1,10 @@
+const express= require('express')
+const {registerUserCtrl, loginUserCtrl, emailVerificationCtrl, resendEmailVerificationCtrl}= require("../../controllers/users/Auth")
+
+
+authRoutes= express.Router()
+
+authRoutes.post('/register', registerUserCtrl);
+
+
+module.exports = authRoutes
